feat: expose middleware utils
