+++ packages/middleware/.babelrc
@@ -1,3 +1,14 @@
 {
-  "extends": "../../.babelrc"
+  "extends": "../../.babelrc",
+  "presets": [
+    [
+      "@babel/env",
+      {
+        "targets": {
+          "node": 12
+        }
+      }
+    ],
+    "@babel/typescript"
+  ]
 }
