+++ packages/middleware/jest.config.js
@@ -1,3 +1,12 @@
 const config = require('../../jest/config');
 
-module.exports = Object.assign({}, config, {});
+module.exports = Object.assign({}, config, {
+  coverageThreshold: {
+    global: {
+      lines: 67,
+      functions: 80,
+      branches: 56,
+      statements: 67,
+    },
+  },
+});
