+++ packages/middleware/package.json
@@ -35,19 +35,23 @@
     "build:types": "tsc --emitDeclarationOnly -p tsconfig.build.json",
     "build:js": "babel src/ --out-dir build/ --copy-files --extensions \".ts,.tsx\" --source-maps",
     "watch": "pnpm build:js -- --watch",
+    "test": "jest",
     "build": "pnpm run build:js && pnpm run build:types"
   },
   "dependencies": {
-    "debug": "4.3.4",
-    "body-parser": "1.20.1",
-    "@verdaccio/auth": "workspace:6.0.0-6-next.34",
     "@verdaccio/core": "workspace:6.0.0-6-next.55",
     "@verdaccio/logger": "workspace:6.0.0-6-next.23",
     "@verdaccio/utils": "workspace:6.0.0-6-next.23",
-    "lodash": "4.17.21"
+    "debug": "4.3.4",
+    "lodash": "4.17.21",
+    "mime": "2.6.0"
   },
   "funding": {
     "type": "opencollective",
     "url": "https://opencollective.com/verdaccio"
+  },
+  "devDependencies": {
+    "body-parser": "1.20.1",
+    "supertest": "6.3.3"
   }
 }
