+++ packages/middleware/src/index.ts
@@ -1,1 +1,17 @@
-export * from './middleware';
+export { match } from './middlewares/match';
+export { setSecurityWebHeaders } from './middlewares/security-headers';
+export { validateName, validatePackage } from './middlewares/validation';
+export { media } from './middlewares/media';
+export { encodeScopePackage } from './middlewares/encode-pkg';
+export { expectJson } from './middlewares/json';
+export { antiLoop } from './middlewares/antiLoop';
+export { final } from './middlewares/final';
+export { allow } from './middlewares/allow';
+export { errorReportingMiddleware, handleError } from './middlewares/error';
+export {
+  log,
+  LOG_STATUS_MESSAGE,
+  LOG_VERDACCIO_BYTES,
+  LOG_VERDACCIO_ERROR,
+} from './middlewares/log';
+export * from './types';
