+++ packages/middleware/src/middlewares/match.ts
@@ -1,0 +1,16 @@
+import { $NextFunctionVer, $RequestExtend, $ResponseExtend } from '../types';
+
+export function match(regexp: RegExp): any {
+  return function (
+    req: $RequestExtend,
+    res: $ResponseExtend,
+    next: $NextFunctionVer,
+    value: string
+  ): void {
+    if (regexp.exec(value)) {
+      next();
+    } else {
+      next('route');
+    }
+  };
+}
