+++ packages/utils/src/index.ts
@@ -3,3 +3,4 @@ export * from './utils';
 export * from './crypto-utils';
 export * from './replace-lodash';
 export * from './matcher';
+export * from './middleware-utils';
